+++ jest.config.js
@@ -1,7 +1,9 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 const path = require('path');
 
-const esModules = ['@folio', '@k-int'].join('|');
+// add '@k-int' for test in FilterFileForm.test.js
+// in DocumentsFieldArray.js we import { useKiwtFieldArray } from '@k-int/stripes-kint-components';
+const esModules = ['@folio', 'ky', '@k-int'].join('|');
 
 module.exports = {
   collectCoverageFrom: [
@@ -13,11 +15,12 @@ module.exports = {
   coverageReporters: ['lcov'],
   reporters: ['jest-junit', 'default'],
   transform: { '^.+\\.(js|jsx)$': path.join(__dirname, './test/jest/jest-transformer.js') },
-  transformIgnorePatterns: [`/node_modules/(?!${esModules}|ky)`],
+  transformIgnorePatterns: [`/node_modules/(?!${esModules})`],
   moduleNameMapper: {
     '^.+\\.(css)$': 'identity-obj-proxy',
     '^.+\\.(svg)$': 'identity-obj-proxy',
   },
+  testEnvironment: 'jsdom',
   testMatch: ['**/(lib|src)/**/?(*.)test.{js,jsx}'],
   testPathIgnorePatterns: ['/node_modules/'],
   setupFiles: [path.join(__dirname, './test/jest/setupTests.js')],