+++ src/components/MetadataSources/MetadataSourceView.test.js
@@ -1,14 +1,11 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import { screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import { StripesContext } from '@folio/stripes/core';
 
-import '../../../test/jest/__mock__';
-import translationsProperties from '../../../test/jest/helpers/translationsProperties';
-import renderWithIntl from '../../../test/jest/helpers';
+import withIntlConfiguration from '../../../test/jest/helpers/withIntlConfiguration';
 import SOURCE from '../../../test/fixtures/metadatasource';
 import MetadataSourceView from './MetadataSourceView';
-// import stripes from '../../../test/jest/__mock__/stripesCore.mock';
 
 const handlers = {
   onClose: jest.fn,
@@ -18,54 +15,68 @@ const handlers = {
 const okapiState = { okapi: { token: {} } };
 
 const stripes = {
-  okapi: { url: '' },
+  // we need to set okapi token here
+  okapi: {
+    tenant: 'diku',
+    token: 'someToken',
+    url: 'https://folio-testing-okapi.dev.folio.org',
+  },
+  // we need to set store here
   store: { getState: () => { return okapiState; } },
 };
 
-const renderMetadataSourceView = (fakeStripes = stripes, record = SOURCE) => (
-  renderWithIntl(
+const renderMetadataSourceView = (record = SOURCE) => (
+  render(withIntlConfiguration(
     <MemoryRouter>
-      <StripesContext.Provider value={fakeStripes}>
+      <StripesContext.Provider value={stripes}>
         <MetadataSourceView
           canEdit
           handlers={handlers}
           isLoading={false}
           record={record}
-          stripes={fakeStripes}
+          stripes={stripes}
         />
       </StripesContext.Provider>
-    </MemoryRouter>,
-    translationsProperties
-  )
+    </MemoryRouter>
+  ))
 );
+
+jest.unmock('react-intl');
 
 describe('MetadataSourceView', () => {
   beforeEach(() => {
-    renderMetadataSourceView(stripes, SOURCE);
+    renderMetadataSourceView(SOURCE);
   });
 
   it('accordions should be present', () => {
     expect(document.querySelector('#managementAccordion')).toBeInTheDocument();
     expect(document.querySelector('#technicalAccordion')).toBeInTheDocument();
   });
+
   it('should display name', () => {
     expect(screen.getByLabelText('Cambridge University Press Journals')).toBeInTheDocument();
   });
+
   it('should display description', () => {
     expect(screen.getByText('Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore.')).toBeInTheDocument();
   });
+
   it('should display status', () => {
     expect(screen.getByText('Implementation')).toBeInTheDocument();
   });
+
   it('should display organization', () => {
     expect(screen.getByText('Test organization')).toBeInTheDocument();
   });
+
   it('should display indexing level', () => {
     expect(screen.getByText('bibliographic')).toBeInTheDocument();
   });
+
   it('should display general notes', () => {
     expect(screen.getByText('Test licensing note')).toBeInTheDocument();
   });
+
   test('should display buttons', async () => {
     expect(screen.getByText('Show selected collections')).toBeInTheDocument();
     expect(screen.getByText('Select all collections')).toBeInTheDocument();