+++ test/bigtest/network/models/finc-select-filter-collection.js
@@ -1,8 +1,0 @@
-import {
-  belongsTo,
-  Model,
-} from 'miragejs';
-
-export default Model.extend({
-  collectionIds: belongsTo('finc-select-filter'),
-});