+++ test/bigtest/network/models/finc-select-filter-organization.js
@@ -1,8 +1,0 @@
-import {
-  belongsTo,
-  Model,
-} from 'miragejs';
-
-export default Model.extend({
-  organization: belongsTo('finc-select-metadata-source'),
-});