+++ test/jest/__mock__/index.js
@@ -1,8 +1,1 @@
-import './currencyData.mock';
-import './documentCreateRange.mock';
-import './matchMedia.mock';
-import './intl.mock';
-import './stripesConfig.mock';
-import './stripesCore.mock';
-import './stripesIcon.mock';
-import './stripesSmartComponents.mock';
+import '@folio/stripes-acq-components/test/jest/__mock__';